+++ core/src/main/java/io/memoria/reactive/core/stream/OMsg.java
@@ -2,4 +2,4 @@ package io.memoria.reactive.core.stream;
 
 import java.io.Serializable;
 
-public record OMsg(int sKey, String value) implements Serializable {}
+public record OMsg(long sKey, String value) implements Serializable {}
