+++ core/src/main/java/io/memoria/reactive/core/stream/OStreamRepo.java
@@ -9,9 +9,9 @@ import reactor.core.publisher.Mono;
 public interface OStreamRepo {
   Mono<String> create(String topic);
 
-  Mono<Integer> publish(String topic, OMsg oMsg);
+  Mono<Long> publish(String topic, OMsg oMsg);
 
-  Mono<Integer> size(String topic);
+  Mono<Long> size(String topic);
 
-  Flux<OMsg> subscribe(String topic, int skipped);
+  Flux<OMsg> subscribe(String topic, long skipped);
 }
