+++ core/src/test/java/io/memoria/reactive/core/stream/mem/OStreamMemRepoTest.java
@@ -13,12 +13,12 @@ import java.time.Duration;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.concurrent.ConcurrentHashMap;
-import java.util.concurrent.atomic.AtomicInteger;
+import java.util.concurrent.atomic.AtomicLong;
 
 public class OStreamMemRepoTest {
-  private static final int N_ELEMENTS = 100000;
+  private static final long N_ELEMENTS = 100000;
   private static final Map<String, Many<OMsg>> db = new ConcurrentHashMap<>();
-  private static final Map<String, AtomicInteger> sizes = new HashMap<>();
+  private static final Map<String, AtomicLong> sizes = new HashMap<>();
   private static final OStreamMemRepo streamRepo = new OStreamMemRepo(db, sizes, 1000);
   private static final String SOME_TOPIC = "NODE_TOPIC";
 
@@ -36,7 +36,7 @@ public class OStreamMemRepoTest {
     // When
     var pub = Flux.fromIterable(msgs).flatMap(msg -> streamRepo.publish(SOME_TOPIC, msg));
     // Then
-    var expected = msgs.map(OMsg::sKey).toJavaArray(Integer[]::new);
+    var expected = msgs.map(OMsg::sKey).toJavaArray(Long[]::new);
     StepVerifier.create(pub).expectNext(expected).verifyComplete();
     StepVerifier.create(streamRepo.size(SOME_TOPIC)).expectNext(N_ELEMENTS).verifyComplete();
   }
@@ -44,11 +44,14 @@ public class OStreamMemRepoTest {
   @Test
   void subscribe() {
     // Given
-    Flux.range(0, N_ELEMENTS).concatMap(this::publish).subscribe();
-    Flux.range(N_ELEMENTS, N_ELEMENTS).concatMap(this::publish).delaySubscription(Duration.ofMillis(10)).subscribe();
+    Flux.range(0, (int) N_ELEMENTS).concatMap(this::publish).subscribe();
+    Flux.range((int) N_ELEMENTS, (int) N_ELEMENTS)
+        .concatMap(this::publish)
+        .delaySubscription(Duration.ofMillis(10))
+        .subscribe();
     // When
     var sub = streamRepo.subscribe(SOME_TOPIC, 0).map(OMsg::sKey).take(2 * N_ELEMENTS);
-    var expected = List.range(0, 2 * N_ELEMENTS).toJavaArray(Integer[]::new);
+    var expected = List.range(0, 2 * N_ELEMENTS).toJavaArray(Long[]::new);
     // Then
     StepVerifier.create(sub).expectNext(expected).verifyComplete();
     StepVerifier.create(streamRepo.size(SOME_TOPIC)).expectNext(2 * N_ELEMENTS).verifyComplete();
@@ -56,7 +59,7 @@ public class OStreamMemRepoTest {
     StepVerifier.create(sub).expectNext(expected).verifyComplete();
   }
 
-  private Mono<Integer> publish(Integer i) {
+  private Mono<Long> publish(long i) {
     return streamRepo.publish(SOME_TOPIC, new OMsg(i, "hello" + i));
   }
 }
